Develop

+++ app/Api/Modules/Client/ClientService.php
@@ -2,7 +2,13 @@
 
 namespace App\Api\Modules\Client;
 
+use App\Api\Modules\Client\Dtos\ClientDto;
 use App\Api\Modules\Client\Entity\ClientEntity;
+use App\Api\Modules\Property\Entity\PropertyEntity;
+use App\Api\Shared\Filters\DatabaseOperation\DatabaseOperationFilter;
+use App\Api\Shared\Services\Mailer\Dtos\SendConversationInviteDto;
+use App\Api\Shared\Services\Mailer\IMailer;
+use App\Api\Shared\Services\Mailer\MailerService;
 use Exception;
 use Raven\Falcon\Http\Response;
 use Raven\Falcon\Http\StatusCode;
@@ -12,18 +18,26 @@ use Raven\Falcon\Http\Exceptions\NotFoundException;
 
 class ClientService
 {
-
-	public function __construct(private readonly ClientEntity $clientEntity = new ClientEntity) {}
+	private MailerService $mailerService;
+	public function __construct(
+		IMailer $iMailer,
+		private readonly ClientEntity $clientEntity = new ClientEntity,
+		private readonly PropertyEntity $propertyEntity = new PropertyEntity,
+		private readonly DatabaseOperationFilter $databaseOperationFilter = new DatabaseOperationFilter
+	) {
+		$this->mailerService = new MailerService($iMailer);
+	}
 
 	public function create(ClientDto $clientDto)
 	{
 		try {
 			$this->clientEntity->create($clientDto);
+
 			return Response::sendBody([
 				"message" => "Cliente criado com sucesso"
 			], StatusCode::CREATED);
 		} catch (Exception $e) {
-			throw new BadRequestException($e->getMessage());
+			$this->databaseOperationFilter->handle($e);
 		}
 	}
 
@@ -31,6 +45,7 @@ class ClientService
 	{
 		try {
 			$clients = $this->clientEntity->selectAll();
+
 			return Response::sendBody($clients);
 		} catch (PDOException $e) {
 			throw new BadRequestException($e->getMessage());
@@ -44,7 +59,7 @@ class ClientService
 
 			if (!$client) throw new BadRequestException('Cliente não encontrado');
 
-			Response::sendBody($client);
+			return Response::sendBody($client);
 		} catch (PDOException $e) {
 			throw new BadRequestException($e->getMessage());
 		}
@@ -54,7 +69,8 @@ class ClientService
 	{
 		try {
 			$this->clientEntity->update($id, $clientDto);
-			Response::sendBody([
+
+			return Response::sendBody([
 				"message" => "Cliente atualizado com sucesso"
 			]);
 		} catch (PDOException $e) {
@@ -64,15 +80,28 @@ class ClientService
 	public function delete(int $id)
 	{
 		try {
+			$this->propertyEntity->selectByClientId($id);
+
+			if ($this->propertyEntity->selectByClientId($id)) throw new BadRequestException("Cliente possui imóveis cadastrados");
+
 			$deletedClient = $this->clientEntity->delete($id);
 
 			if (!$deletedClient) throw new NotFoundException("Cliente não encontrado");
 
-			Response::sendBody([
+			return Response::sendBody([
 				"message" => "Cliente removido com sucesso"
 			]);
 		} catch (PDOException $e) {
 			throw new BadRequestException($e->getMessage());
 		}
+	}
+
+	public function sendConversationInviteEmail(SendConversationInviteDto $sendConversationInviteDto)
+	{
+		$this->mailerService->sendConversationInvite(getenv("TALK_TO_US_EMAIL"), $sendConversationInviteDto);
+
+		return Response::sendBody([
+			"message" => "Email de contato enviado com sucesso"
+		]);
 	}
 }
